+++ pontos/models.py
@@ -2,24 +2,7 @@ from uuid import uuid4
 from django.db import models
 from atracoes.models import Atracao
 from avaliacoes.models import Avaliacao
-
-
-class Endereco(models.Model):
-    codigo = models.UUIDField(
-        default=uuid4,
-        editable=False,
-        primary_key=True,
-        db_column='cd_endereco'
-    ),
-    latitude = models.CharField(
-        max_length=20,
-        null=True
-    )
-    longitude = models.CharField(
-        max_length=20,
-        null=True
-    )
-    descricao = models.TextField()
+from comum.models import Endereco
 
 
 class PontoTuristico(models.Model):
